Refactor database models to function same as other models.
