+++ models/context.test.js
@@ -3,7 +3,7 @@
 const path = require('path')
 const fs = require('fs')
 const fixture = fs.readFileSync(path.join(__fixturesDir, 'context.json'), 'utf-8')
-const {Context} = require('./context')
+const Context = require('./context')
 
 describe('Context', () => {
   it('should be defined', () => {
