+++ test/global-teardown.js
@@ -1,8 +1,13 @@
 'use strict'
 
-const {sequelize} = require('./factories')
+const sequelize = require('../config/sequelize')
+const {forEach} = require('lodash')
 
 module.exports = () => {
-  // Close any remaining database connections
-  return sequelize.close()
+  // Destroy all Model instances (wipe the database) then close database connections
+  const destroyed = []
+  forEach(sequelize.models || [], (value) => {
+    destroyed.push(value.destroy({truncate: true}))
+  })
+  return Promise.all(destroyed).then(() => sequelize.close())
 }
